pipeline/requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
openpyxl>=3.1.0
dnspython>=2.4.0
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
asyncio
